Split implementation of different split methods in separated files, and move all of them to a separated directory, so they can be used by other tree types.
