+++ src/mlpack/core/tree/space_split/space_split.hpp
@@ -2,9 +2,8 @@
  * @file space_split.hpp
  * @author Marcos Pividori
  *
- * Definition of MidpointSpaceSplit and MeanSpaceSplit, to create a splitting
- * hyperplane considering the midpoint/mean of the values in a certain
- * projection.
+ * Definition of SpaceSplit, implementing some methods to create a projection
+ * vector based on a given set of points.
  */
 #ifndef MLPACK_CORE_TREE_SPILL_TREE_SPACE_SPLIT_HPP
 #define MLPACK_CORE_TREE_SPILL_TREE_SPACE_SPLIT_HPP
@@ -14,50 +13,6 @@
 
 namespace mlpack {
 namespace tree {
-
-template<typename MetricType, typename MatType>
-class MeanSpaceSplit
-{
- public:
-  /**
-   * Create a splitting hyperplane considering the mean of the values in a
-   * certain projection.
-   *
-   * @param bound The bound used for this node.
-   * @param data The dataset used by the tree.
-   * @param points Vector of indexes of points to be considered.
-   * @param hyp Resulting splitting hyperplane.
-   * @return Flag to determine if split is possible.
-   */
-  template<typename HyperplaneType>
-  static bool SplitSpace(
-      const typename HyperplaneType::BoundType& bound,
-      const MatType& data,
-      const arma::Col<size_t>& points,
-      HyperplaneType& hyp);
-};
-
-template<typename MetricType, typename MatType>
-class MidpointSpaceSplit
-{
- public:
-  /**
-   * Create a splitting hyperplane considering the midpoint of the values in a
-   * certain projection.
-   *
-   * @param bound The bound used for this node.
-   * @param data The dataset used by the tree.
-   * @param points Vector of indexes of points to be considered.
-   * @param hyp Resulting splitting hyperplane.
-   * @return Flag to determine if split is possible.
-   */
-  template<typename HyperplaneType>
-  static bool SplitSpace(
-      const typename HyperplaneType::BoundType& bound,
-      const MatType& data,
-      const arma::Col<size_t>& points,
-      HyperplaneType& hyp);
-};
 
 template<typename MetricType, typename MatType>
 class SpaceSplit
