+++ src/mlpack/core/tree/space_split/space_split_impl.hpp
@@ -2,9 +2,8 @@
  * @file space_split_impl.hpp
  * @author Marcos Pividori
  *
- * Implementation of MidpointSpaceSplit and MeanSpaceSplit, to create a
- * splitting hyperplane considering the midpoint/mean of the values in a certain
- * projection.
+ * Implementation of SpaceSplit, to create a projection vector based on a given
+ * set of points.
  */
 #ifndef MLPACK_CORE_TREE_SPILL_TREE_SPACE_SPLIT_IMPL_HPP
 #define MLPACK_CORE_TREE_SPILL_TREE_SPACE_SPLIT_IMPL_HPP
@@ -13,51 +12,6 @@
 
 namespace mlpack {
 namespace tree {
-
-template<typename MetricType, typename MatType>
-template<typename HyperplaneType>
-bool MeanSpaceSplit<MetricType, MatType>::SplitSpace(
-    const typename HyperplaneType::BoundType& bound,
-    const MatType& data,
-    const arma::Col<size_t>& points,
-    HyperplaneType& hyp)
-{
-  typename HyperplaneType::ProjVectorType projVector;
-  double midValue;
-
-  if (!SpaceSplit<MetricType, MatType>::GetProjVector(bound, data, points,
-    projVector, midValue))
-    return false;
-
-  double splitVal = 0.0;
-  for (size_t i = 0; i < points.n_elem; i++)
-    splitVal += projVector.Project(data.col(points[i]));
-  splitVal /= points.n_elem;
-
-  hyp = HyperplaneType(projVector, splitVal);
-
-  return true;
-}
-
-template<typename MetricType, typename MatType>
-template<typename HyperplaneType>
-bool MidpointSpaceSplit<MetricType, MatType>::SplitSpace(
-    const typename HyperplaneType::BoundType& bound,
-    const MatType& data,
-    const arma::Col<size_t>& points,
-    HyperplaneType& hyp)
-{
-  typename HyperplaneType::ProjVectorType projVector;
-  double midValue;
-
-  if (!SpaceSplit<MetricType, MatType>::GetProjVector(bound, data, points,
-    projVector, midValue))
-    return false;
-
-  hyp = HyperplaneType(projVector, midValue);
-
-  return true;
-}
 
 template<typename MetricType, typename MatType>
 bool SpaceSplit<MetricType, MatType>::GetProjVector(
